Add an api to get and set a gc blacklist, to schedule nodes for removal from DDFS.
The gc blacklist differs from a write blacklist in the following sense: a node
on the write_blacklist will not have new DDFS data being written to it, but any
existing data is treated as stable.  A node on the gc blacklist is treated as
unstable; additional replicas are generated for any data and metadata on that
node.