Add a test to distinguish between group_node_label and group_label.
When we run this test on a disco cluster with more than one node, we can
see the improvements of the group_node_lable over group_label.  With
group_node_label, for each intermediate step, one task is created on each
node and the inputs are processed locally.  With the later, only one
task is created on one of the nodes and all of the input is shipped to
that node.  If the size of the input files is large enough, we can see
significant improvements.