Make the job coordinator a system process.
The reason the job coordinator was not a system process before was that
it's terminate would have been called when its parent (the http request
handler) finished.  This is not an erlang bug, the behaviour is explained
in this mail thread:
http://erlang.org/pipermail/erlang-questions/2005-January/013943.html
and in this stack overflow answer:
http://stackoverflow.com/a/7753788/830681

The solution is linking the http handler from the newly created child.