redis scheme: Use a list of values instead of a single value.
There might be some duplicates in the (key, value) pairs and we will need all of
those duplicates (not jsut one of them).  Therefore, we have to use a list.

In some applications, a single value for each key might be enough.  The list of
values works for those applications too.