worker: Increase the worker protocol timeout.
Disco might have to prepare the inputs for the tasks and it certianly needs more
than 10 minutes to do that.  For concurrent stages, for some type of groupings,
Disco needs all of the tasks of the previous stage to finish.

The current timeout (3 days) is just a large value.  If any worker failed with a
timeout, we have to increase this value.