disco_worker: use the update function instead of directly sending inputs to the worker.
When the worker_runtime decides that it should not respond to an input message
immediately, the disco_worker can avoid consuming that INPUT message from the
worker and keep it pending.

That means, when the input is ready, disco_worker can just call the update
functio.  The INPUT message is going to be consumed but this time, there are
some inputs available which will bee sent to the worker.