event_server: use processes for storing job events.
Instead of storing the events of the jobs in the status of the event_server, we
create one job_event process for each job and delegate the responsibility of
handling the job events to these processes. All of the operations still go
through the event_server, therefore, it is still a bottleneck.